adjusted files not directly related to tests
adjusted test helpers

The TLS helper was adjusted to the new helper scripts; the default setup
helper saw a bugix (expanding the array properly) and the addition of
the IPv4 defaults. The IPv4 default was added after recent Docker pushes
IPv6 and I saw weird IPv6 related errors in the logs.. now we're sure
IPv4 is used.