added functionality to check whether program is running
This change adds a helper function to check whether a program is running
inside a container or not. A side product is a function like
`_run_in_container` but with an explicit contaner name. Future PRs can
use this helper function now to check whether a process is running or
not.

This was done for the tests of Fail2Ban, but can be used for other tests
in the future as well.